+++ src/ui/hooks/useRecipies.ts
@@ -1,13 +1,13 @@
 import { useContext, useEffect, useState } from "react";
 import { DBContext } from "../providers/database/db-context";
 import { ActionQueue } from "../../utils/action-queue";
-import { Recipie } from "../../models/recipies";
+import { Recipe } from "../../models/recipies";
 
 const queue = new ActionQueue();
 
-export function useRecipies(): [Recipie[], (recipie: Recipie) => void] {
+export function useRecipies(): [Recipe[], (recipe: Recipe) => void] {
   const { stores } = useContext(DBContext);
-  const [recipies, setRecipies] = useState<Recipie[]>([]);
+  const [recipies, setRecipies] = useState<Recipe[]>([]);
 
   useEffect(() => {
     if (!stores) return;
@@ -18,13 +18,13 @@ export function useRecipies(): [Recipie[], (recipie: Recipie) => void] {
     })();
   }, [stores]);
 
-  function deleteRecipie(recipie: Recipie) {
+  function deleteRecipie(recipe: Recipe) {
     if (!stores) return;
 
     queue.enqueue(async () => {
-      await stores.recipieStore.delete(recipie.id);
+      await stores.recipieStore.delete(recipe.id);
     });
-    setRecipies(recipies.filter(i => i.id !== recipie.id));
+    setRecipies(recipies.filter(i => i.id !== recipe.id));
   }
 
   return [recipies, deleteRecipie];