+++ docs/CHANGELOG.md
@@ -50,9 +50,9 @@ Bugs:
 
 Bugs:
 
-- [#35](https://github.com/benjamin-wright/meal-planner/issues/35) - Recipie page not scrolling
+- [#35](https://github.com/benjamin-wright/meal-planner/issues/35) - Recipe page not scrolling
 - [#33](https://github.com/benjamin-wright/meal-planner/issues/33) - factoring out a common component
-- [#31](https://github.com/benjamin-wright/meal-planner/issues/31) - Cancelling ingredient clears the recipie
+- [#31](https://github.com/benjamin-wright/meal-planner/issues/31) - Cancelling ingredient clears the recipe
 
 ## Version 1.0.6